+++ manifests/cf-manifest/spec/integration/integration_spec.rb
@@ -40,34 +40,16 @@ RSpec.describe "generic manifest validations" do
   end
 
   describe "name uniqueness" do
-    context "cloud-config" do
-      %w(
-        disk_types
-        networks
-        vm_types
-      ).each do |resource_type|
-        specify "all #{resource_type} have a unique name" do
-          all_resource_names = cloud_config_with_defaults.fetch(resource_type).map { |r| r["name"] }
+    %w(
+      instance_groups
+      releases
+    ).each do |resource_type|
+      specify "all #{resource_type} have a unique name" do
+        all_resource_names = manifest.fetch(resource_type).map { |r| r["name"] }
 
-          duplicated_names = all_resource_names.select { |n| all_resource_names.count(n) > 1 }.uniq
-          expect(duplicated_names).to be_empty,
-            "found duplicate names (#{duplicated_names.join(',')}) for #{resource_type}"
-        end
-      end
-    end
-
-    context "manifest" do
-      %w(
-        instance_groups
-        releases
-      ).each do |resource_type|
-        specify "all #{resource_type} have a unique name" do
-          all_resource_names = manifest.fetch(resource_type).map { |r| r["name"] }
-
-          duplicated_names = all_resource_names.select { |n| all_resource_names.count(n) > 1 }.uniq
-          expect(duplicated_names).to be_empty,
-            "found duplicate names (#{duplicated_names.join(',')}) for #{resource_type}"
-        end
+        duplicated_names = all_resource_names.select { |n| all_resource_names.count(n) > 1 }.uniq
+        expect(duplicated_names).to be_empty,
+          "found duplicate names (#{duplicated_names.join(',')}) for #{resource_type}"
       end
     end
   end
