+++ manifests/cloud-config/.rspec
@@ -1,0 +1,2 @@
+--color
+--require spec_helper
