+++ manifests/cloud-config/operations.d/060-cdn-broker.yml
