+++ manifests/cloud-config/operations.d/070-elasticache-broker.yml
