+++ manifests/cloud-config/operations.d/090-add-bosh-client.yml
