+++ manifests/cloud-config/scripts/generate-cloud-config.sh
@@ -6,7 +6,7 @@ PAAS_CF_DIR=${PAAS_CF_DIR:-paas-cf}
 WORKDIR=${WORKDIR:-.}
 
 opsfile_args=""
-for i in "${PAAS_CF_DIR}"/manifests/cf-manifest/cloud-config/operations.d/*.yml; do
+for i in "${PAAS_CF_DIR}"/manifests/cloud-config/operations.d/*.yml; do
   opsfile_args+="-o $i "
 done
 
@@ -15,11 +15,8 @@ bosh interpolate \
   --var-errs \
   --vars-file="${WORKDIR}/terraform-outputs/vpc.yml" \
   --vars-file="${WORKDIR}/terraform-outputs/bosh.yml" \
-  --vars-file="${WORKDIR}/terraform-outputs/concourse.yml" \
   --vars-file="${WORKDIR}/terraform-outputs/cf.yml" \
   --vars-file="${WORKDIR}/cf-secrets/cf-secrets.yml" \
   --vars-file="${PAAS_CF_DIR}/manifests/variables.yml" \
-  --vars-file="${ENV_SPECIFIC_BOSH_VARS_FILE}" \
   ${opsfile_args} \
-  "$@" \
-  "${PAAS_CF_DIR}/manifests/cf-manifest/cloud-config/000-base-cloud-config.yml"
+  "${PAAS_CF_DIR}/manifests/cloud-config/cloud-config.yml"
