+++ manifests/cloud-config/spec/compilation_spec.rb
