+++ manifests/cloud-config/spec/networks_spec.rb
