+++ manifests/cloud-config/spec/spec_helper.rb
@@ -117,5 +117,4 @@ end
 # option on the command line or in ~/.rspec, .rspec or `.rspec-local`.
 #
 Dir[File.expand_path("../../shared/spec/support/**/*.rb", __dir__)].each { |f| require f }
-Dir[File.expand_path("../../cloud-config/spec/support/**/*.rb", __dir__)].each { |f| require f }
 Dir[File.expand_path("support/**/*.rb", __dir__)].each { |f| require f }
