+++ manifests/cloud-config/spec/vm_extensions_spec.rb
