+++ manifests/cloud-config/spec/vm_types_spec.rb
